StoneChallenge.CostManaging (stone_challenge v0.1.0)

Interface for interacting with cost managing modules.

Functions

create_purchase_item(attrs)

Specs

create_purchase_item(%{
  name: String.t(),
  quantity: integer(),
  unit_price_in_cents: integer()
}) :: {:error, String.t()} | {:ok, StoneChallenge.CostManaging.PurchaseItem.t()}

Returns a struct of StoneChallenge.CostManaging.PurchaseItem type.

Parameters

attrs: map with name as string, quantity as integer and unit_price_in_cents as integer.

divide_costs_among_emails(purchase_list, emails_list)

Specs

divide_costs_among_emails([StoneChallenge.CostManaging.PurchaseItem.t()], [
  String.t()
]) :: {:error, String.t()} | {:ok, %{required(String.t()) => integer()}}

Returns a map of String => integer structure, representing the divided price to be paid by each 'email'.

Parameters

purchase_list: list of StoneChallenge.CostManaging.PurchaseItem structs

emails: list of strings