StoneChallenge.DataGeneration (stone_challenge v0.1.0)

Interface for interacting with data generating modules.

Functions

email_list(ammount_of_emails \\ Enum.random(1..5))

Specs

email_list(integer()) :: [String.t()]

Returns a list of strings representing emails.

Parameters

ammount_of_emails: optional parameter that sets the ammount of emails to be present in the list.

purchase_list(ammount_of_items \\ Enum.random(1..10))

Specs

purchase_list(integer()) :: [StoneChallenge.CostManaging.PurchaseItem.t()]

Returns a list of StoneChallenge.CostManaging.PurchaseItem structs.

Parameters

ammount_of_items: optional parameter that sets the ammount of structs to be present in the list.