StoneChallenge.DataGeneration.Lists (stone_challenge v0.1.0)

Provides functions for lists of unique email strings or PurchaseItem structs.

Functions

emails(ammount_of_emails)

Specs

emails(integer()) :: [String.t()]

Generates a list of numberend, and thus unique, strings representing emails.

Parameters

ammount_of_emails: Integer with number of emails to be returned in the list.

Examples

iex> StoneChallenge.DataGeneration.Lists.emails(2) ["estevan1@amail.com", "vinicius2@bmail.com"]

purchase(ammount_of_items)

Specs

purchase(integer()) :: [StoneChallenge.CostManaging.PurchaseItem.t()]

Generates a list of StoneChallenge.CostManaging.PurchaseItem structs.

Parameters

ammount_of_items: Integer with number of items to be returned in the list.

Example

iex> StoneChallenge.DataGeneration.Lists.purchase(3) [

%StoneChallenge.CostManaging.PurchaseItem{
  name: "Mango",
  quantity: 2,
  unit_price_in_cents: 50
},
%StoneChallenge.CostManaging.PurchaseItem{
  name: "Banana",
  quantity: 2,
  unit_price_in_cents: 360
},
%StoneChallenge.CostManaging.PurchaseItem{
  name: "Pineapple",
  quantity: 5,
  unit_price_in_cents: 3900
}

]