StoneChallenge.DataGeneration.Lists (stone_challenge v0.1.0)

Provides functions for lists of unique email strings or PurchaseItem structs.

Functions

type(type, ammount)

Specs

type(:email | :purchase_item, integer()) :: [
  String.t() | StoneChallenge.CostManaging.PurchaseItem.t()
]

Generates a list of specified type.

Parameters

type: atom representing desired type of list elements.

ammount: integer with number of elements of specified type in list.

Examples

  iex> StoneChallenge.DataGeneration.Lists.type(:email, 2)

  ["estevan1@amail.com", "vinicius2@bmail.com"]

  iex> StoneChallenge.DataGeneration.Lists.type(:purchase_item, 3)

  [
    %StoneChallenge.CostManaging.PurchaseItem{
      name: "Mango",
      quantity: 2,
      unit_price_in_cents: 50
    },
    %StoneChallenge.CostManaging.PurchaseItem{
      name: "Banana",
      quantity: 2,
      unit_price_in_cents: 360
    },
    %StoneChallenge.CostManaging.PurchaseItem{
      name: "Pineapple",
      quantity: 5,
      unit_price_in_cents: 3900
    }
  ]